Use size_t

+++ include/envoy/formatter/substitution_formatter.h
@@ -96,7 +96,7 @@ public:
    * @return FormattterProviderPtr substitution provider for the parsed command
    */
   virtual FormatterProviderPtr parse(const std::string& token, size_t pos,
-                                     int command_end_position) const PURE;
+                                     size_t command_end_position) const PURE;
 };
 
 using CommandParserPtr = std::unique_ptr<CommandParser>;
